python/requirements.txt
absl-py==1.2.0
astunparse==1.6.3
blis==0.9.1
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
confection==0.0.1
cymem==2.0.6
emoji==2.0.0
filelock==3.8.0
Flask==2.2.2
flatbuffers==1.12
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.1
h5py==3.7.0
huggingface-hub==0.9.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
jarowinkler==1.2.1
Jinja2==3.1.2
keras==2.9.0
Keras-Preprocessing==1.1.2
langcodes==3.3.0
Levenshtein==0.20.3
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
murmurhash==1.0.8
numpy==1.23.3
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
pathy==0.6.2
preshed==3.0.7
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
rapidfuzz==2.6.1
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
ro-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/ro_core_news_lg-3.4.0/ro_core_news_lg-3.4.0-py3-none-any.whl
roner==1.0.3
rsa==4.9
six==1.16.0
smart-open==5.2.1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
stanza==1.4.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-macos==2.9.2
termcolor==1.1.0
thinc==8.1.1
tokenizers==0.12.1
torch==1.12.1
tqdm==4.64.1
transformers==4.21.3
typer==0.4.2
typing_extensions==4.3.0
Unidecode==1.3.4
urllib3==1.26.12
wasabi==0.10.1
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
